Restore ring_buffer code in the  right place: ring_buffer.[h/c], add also macro for STATIC initialization using C++11
